+++ src/editorial/EditorialPages.tsx
@@ -1,3 +1,4 @@
+import { ReferencesPage } from "./ReferencesPage";
 import { HomePage, PartIPage, ResearchPage } from "./ProgrammePages";
 import { PartIIIPage, PartIVPage } from "./EvidencePages";
 import { AboutPage, ExplainersPage, OutputsPage, PaperPage, UpdatesPage } from "./ResourcePages";
@@ -8,6 +9,7 @@ export function renderEditorialPage(path: string) {
     case "research/part-i": return <PartIPage />;
     case "research/part-iii": return <PartIIIPage />;
     case "research/part-iv": return <PartIVPage />;
+    case "references": return <ReferencesPage />;
     case "paper": return <PaperPage />;
     case "outputs": return <OutputsPage />;
     case "explainers": return <ExplainersPage />;